Refactor of package-names, refactor of LifePart
I refactored life part and some package names to have the same naming convention and structure in the project
