+++ AsteroidsEntityFramework/Asteroid/src/main/resources/META-INF/services/dk.sdu.mmmi.cbse.common.services.IGamePluginService
@@ -1,1 +1,1 @@
-dk.sdu.mmmi.cbse.asteroid.AsteroidPlugin
+dk.sdu.mmmi.cbse.asteroidSystem.AsteroidPlugin
