+++ AsteroidsEntityFramework/Bullet/src/main/java/dk/sdu/mmmi/cbse/bulletSystem/BulletControlSystem.java
@@ -1,4 +1,4 @@
-package dk.sdu.mmmi.cbse;
+package dk.sdu.mmmi.cbse.bulletSystem;
 
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
