+++ AsteroidsEntityFramework/Bullet/src/main/java/dk/sdu/mmmi/cbse/bulletSystem/BulletPlugin.java
@@ -1,13 +1,11 @@
-package dk.sdu.mmmi.cbse;
+package dk.sdu.mmmi.cbse.bulletSystem;
 
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
-import dk.sdu.mmmi.cbse.common.data.World;
 import dk.sdu.mmmi.cbse.common.data.entityparts.LifePart;
 import dk.sdu.mmmi.cbse.common.data.entityparts.MovingPart;
 import dk.sdu.mmmi.cbse.common.data.entityparts.PositionPart;
 import dk.sdu.mmmi.cbse.common.services.IBulletPluginService;
-import dk.sdu.mmmi.cbse.common.services.IGamePluginService;
 
 public class BulletPlugin implements IBulletPluginService {
