+++ AsteroidsEntityFramework/Bullet/src/main/resources/META-INF/services/dk.sdu.mmmi.cbse.common.services.IEntityProcessingService
@@ -1,2 +1,2 @@
-dk.sdu.mmmi.cbse.BulletControlSystem
+dk.sdu.mmmi.cbse.bulletSystem.BulletControlSystem
