+++ AsteroidsEntityFramework/Collision/src/main/java/dk/sdu/mmmi/cbse/collisionSystem/CollisionSystem.java
@@ -1,4 +1,4 @@
-package dk.sdu.mmmi.cbse;
+package dk.sdu.mmmi.cbse.collisionSystem;
 
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
