+++ AsteroidsEntityFramework/Collision/src/main/resources/META-INF/services/dk.sdu.mmmi.cbse.common.services.IPostEntityProcessingService
@@ -1,1 +1,1 @@
-dk.sdu.mmmi.cbse.CollisionSystem
+dk.sdu.mmmi.cbse.collisionSystem.CollisionSystem
