+++ AsteroidsEntityFramework/Core/src/main/java/dk/sdu/mmmi/cbse/main/Game.java
@@ -5,24 +5,21 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
-import com.badlogic.gdx.math.MathUtils;
-import dk.sdu.mmmi.cbse.BulletControlSystem;
-import dk.sdu.mmmi.cbse.BulletPlugin;
-import dk.sdu.mmmi.cbse.CollisionSystem;
-import dk.sdu.mmmi.cbse.asteroid.AsteroidControlSystem;
-import dk.sdu.mmmi.cbse.asteroid.AsteroidPlugin;
+import dk.sdu.mmmi.cbse.bulletSystem.BulletControlSystem;
+import dk.sdu.mmmi.cbse.collisionSystem.CollisionSystem;
+import dk.sdu.mmmi.cbse.asteroidSystem.AsteroidControlSystem;
+import dk.sdu.mmmi.cbse.asteroidSystem.AsteroidPlugin;
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
 import dk.sdu.mmmi.cbse.common.data.World;
-import dk.sdu.mmmi.cbse.common.services.IBulletPluginService;
 import dk.sdu.mmmi.cbse.common.services.IEntityProcessingService;
 import dk.sdu.mmmi.cbse.common.services.IGamePluginService;
 import dk.sdu.mmmi.cbse.common.services.IPostEntityProcessingService;
-import dk.sdu.mmmi.cbse.enemysystem.EnemyControlSystem;
-import dk.sdu.mmmi.cbse.enemysystem.EnemyPlugin;
+import dk.sdu.mmmi.cbse.enemySystem.EnemyControlSystem;
+import dk.sdu.mmmi.cbse.enemySystem.EnemyPlugin;
 import dk.sdu.mmmi.cbse.managers.GameInputProcessor;
-import dk.sdu.mmmi.cbse.playersystem.PlayerControlSystem;
-import dk.sdu.mmmi.cbse.playersystem.PlayerPlugin;
+import dk.sdu.mmmi.cbse.playerSystem.PlayerControlSystem;
+import dk.sdu.mmmi.cbse.playerSystem.PlayerPlugin;
 import java.util.ArrayList;
 import java.util.List;
