+++ AsteroidsEntityFramework/Player/src/main/java/dk/sdu/mmmi/cbse/playerSystem/PlayerPlugin.java
@@ -1,6 +1,5 @@
-package dk.sdu.mmmi.cbse.playersystem;
+package dk.sdu.mmmi.cbse.playerSystem;
 
-import dk.sdu.mmmi.cbse.BulletPlugin;
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
 import dk.sdu.mmmi.cbse.common.data.World;
