+++ AsteroidsEntityFramework/Player/src/main/resources/META-INF/services/dk.sdu.mmmi.cbse.common.services.IEntityProcessingService
@@ -1,1 +1,1 @@
-dk.sdu.mmmi.cbse.playersystem.PlayerControlSystem
+dk.sdu.mmmi.cbse.playerSystem.PlayerControlSystem
