+++ AsteroidsEntityFramework/pom.xml
@@ -13,6 +13,7 @@
         <module>Asteroid</module>
         <module>Collision</module>
         <module>Bullet</module>
+        <module>CommonAsteroid</module>
     </modules>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
